Refactor handleRequest to just be in the worker
